// src/app/features/blog-post/models/add-blog-post.model.ts
import { Category } from "../../category/models/category.model";

export interface AddBlogPost {
    title: string;
    shortDescription: string;
    content: string;
    featuredImageUrl: string;
    urlHandle: string;
    author: string;
    publishedDate: Date;
    isVisible: boolean;
    categories: string[];
}